+++ index.html
@@ -1,3 +1,4 @@
+
 <!DOCTYPE html>
 <html lang="hu">
   <head>
@@ -67,4 +68,46 @@
     <meta name="description" content="Aszteroida-vészjelző rendszer" />
     <meta name="author" content="https://nasa-ab2f1.web.app" />
     <link rel="preload" href="/fonts/my-custom-font.woff2" as="font" type="font/woff2" crossorigin="anonymous">
+
+
+
+    <!DOCTYPE html>
+<html lang="hu">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Aszteroida-vészjelző rendszer</title>
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title> Aszteroida-vészjelző rendszer</title>
+    <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
+    <link rel="alternate" hreflang="hu" href="https://nasa-ab2f1.web.app/hu/" />
+    <link rel="alternate" hreflang="en" href="https://nasa-ab2f1.web.app/en/" />
+    <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
+    <meta property="og:title" content="Nasa interface" />
+    <meta property="og:type" content="website" />
+    <meta
+      property="og:description"
+      content="Aszteroida vészjelző rendszer. NASA adatok alapján figyeljük a Földet fenyegető aszteroidákat!"
+    />
+    <meta property="og:url" content="https://nasa-ab2f1.web.app" />
+    <meta property="og:image:type" content="image/jpeg" />
+    <meta
+      property="og:image"
+      content="https://nasa-ab2f1.web.app/opengraphnak.JPG"
+    />
+
+    <meta property="og:image:width" content="820" />
+    <meta property="og:image:height" content="360" />
+    <meta property="og:image:alt" content="Home Page Image" />
+    <meta name="keywords" content="Nasa interface" />
+    <meta name="description" content="Aszteroida-vészjelző rendszer" />
+    <meta name="author" content="https://nasa-ab2f1.web.app" />
+    <link rel="preload" href="/fonts/my-custom-font.woff2" as="font" type="font/woff2" crossorigin="anonymous">
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="module" src="/src/main.jsx"></script>
+  </body>
+</html>
  -->
