+++ package.json
@@ -8,8 +8,8 @@
     "build": "vite build",
     "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
-   "serve": "serve -s build",
-   "diff": "git diff > logs/diff.patch"
+    "serve": "serve -s build",
+    "diff": "git diff > logs/diff.patch"
   },
   "dependencies": {
     "chart.js": "^4.4.2",
