+++ vite.config.js
@@ -1,29 +1,10 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react-swc'
-
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react-swc';
 export default defineConfig({
   plugins: [react()],
+  assetsInclude: ['**/*.JPG', '**/*.PNG'], // Hozzáadtuk a .PNG-t
+  
   build: {
-    rollupOptions: {
-      input: {
-        main: 'src/main.jsx'
-      },
-      output: {
-        assetFileNames: (assetInfo) => {
-          let extType = assetInfo.name.split('.').pop();
-          if (/js/i.test(extType)) {
-            return `assets/[name]-[hash][extname]`;
-          }
-          if (/png|jpe?g|svg|gif|tiff|bmp|ico/i.test(extType)) {
-            extType = 'img';
-          }
-          return `assets/${extType}/[name]-[hash][extname]`;
-        },
-        chunkFileNames: 'assets/[name]-[hash].js',
-        entryFileNames: 'assets/[name]-[hash].js',
-      },
-    },
+    outDir: 'build', // Itt módosítottuk a build mappa nevét.
   },
-})
-
-
+});
