+++ vite_mentes.js
@@ -49,3 +49,64 @@ export default defineConfig({
     },
   },
 })
+
+// Működik csak dist
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react-swc'
+
+export default defineConfig({
+  plugins: [react()],
+  build: {
+    rollupOptions: {
+      input: {
+        main: 'src/main.jsx'
+      },
+      output: {
+        assetFileNames: (assetInfo) => {
+          let extType = assetInfo.name.split('.').pop();
+          if (/js/i.test(extType)) {
+            return `assets/[name]-[hash][extname]`;
+          }
+          if (/png|jpe?g|svg|gif|tiff|bmp|ico/i.test(extType)) {
+            extType = 'img';
+          }
+          return `assets/${extType}/[name]-[hash][extname]`;
+        },
+        chunkFileNames: 'assets/[name]-[hash].js',
+        entryFileNames: 'assets/[name]-[hash].js',
+      },
+    },
+  },
+})
+
+
+nem működik lehet tailwind
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react-swc'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [react()],
+  build: {
+    outDir: 'build', // Add this line
+    rollupOptions: {
+      input: {
+        main: 'src/main.jsx'
+      },
+      output: {
+        assetFileNames: (assetInfo) => {
+          let extType = assetInfo.name.split('.').pop();
+          if (/js|css/i.test(extType)){
+            return `assets/[name]-[hash][extname]`;
+          }
+          if (/png|jpe?g|svg|gif|tiff|bmp|ico/i.test(extType)) {
+            extType = 'img';
+          }
+          return `assets/${extType}/[name]-[hash][extname]`;
+        },
+        chunkFileNames: 'assets/[name]-[hash].js',
+        entryFileNames: 'assets/[name]-[hash].js',
+      }
+    },
+  },
+})
